getIdsOfChosenAnswers
