+++ src/selectors/questionnaire/select_ids_of_chosen_questions.ts
@@ -1,7 +1,7 @@
 import { Store } from '../../stores';
-import { Id, AnswersMap } from '../../stores/questionnaire';
+import { Id } from '../../stores/questionnaire';
 import { toValidOrThrow, ValidStore as QuestionnaireStore } from '../../stores/questionnaire/stores';
-import * as R from 'ramda';
+import { getIdsOfChosenAnswers } from './get_ids_of_chosen_answers';
 
 export const selectIdsOfChosenAnswers = (appStore: Store): ReadonlyArray<Id> => (
     getIdsOfChosenAnswers(pickQuestionnaire(appStore).answers)
@@ -9,16 +9,4 @@ export const selectIdsOfChosenAnswers = (appStore: Store): ReadonlyArray<Id> => 
 
 const pickQuestionnaire = (appStore: Store): QuestionnaireStore => (
     toValidOrThrow(appStore.questionnaireInStore)
-);
-
-export const getIdsOfChosenAnswers = (answers: AnswersMap): ReadonlyArray<Id> => (
-    getIds(filterChosenAnswers(answers))
-);
-
-const getIds = (answers: AnswersMap): ReadonlyArray<Id> => (
-    R.map(R.prop('id'), R.values(answers))
-);
-
-const filterChosenAnswers = (answers: AnswersMap): AnswersMap => (
-    R.filter(R.propEq('isChosen', true), answers)
 );
